requirements.txt
revolt.py
psutil
python-dotenv
filetype